refactor(log): one implementation of the loader-failure line
