+++ src/cli/commands/tool.ts
@@ -5,7 +5,7 @@ import { loadDotEnv } from "../../env.ts";
 import { loadConfig } from "../../engines/pi/config.ts";
 
 import { resolveAgentTools } from "../../engines/pi/create.ts";
-import { reportModuleLoadFailures } from "../../engines/pi/report.ts";
+import { reportModuleLoadFailures } from "../../log.ts";
 import { turnContext } from "../../engines/pi/tool-context.ts";
 import { failStartup, failUsage, placementOrExit } from "../fail.ts";
